Include themify fonts in build
Move Themify fonts to static to workaround the way Hugo handles assets.

Since the fonts are only referenced directly in the themify css and not
via a Permalink tag they weren't being copied to the public folder.

The downside of this approach is they aren't being fingerprinted.
Hopefully this can be improved, but for now we just need to fix the
broken links.